requirements.txt
# ============================================
# NBA STATS API - Core Dependencies
# ============================================

# FastAPI Framework & Web Server
fastapi==0.123.9
uvicorn[standard]==0.38.0
starlette==0.50.0

# Database & ORM
SQLAlchemy==2.0.44
asyncpg==0.31.0
psycopg2-binary==2.9.9

# Environment & Configuration
python-dotenv==1.1.1

# Data Validation & Serialization
pydantic==2.12.4
pydantic_core==2.41.5
email-validator==2.3.0

# Web Scraping
beautifulsoup4==4.13.3
lxml==5.3.2

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib==1.7.4
argon2-cffi==23.1.0
cryptography==46.0.3

# HTTP & Async
httpx==0.28.1
httpcore==1.0.9
requests==2.32.3
anyio==4.11.0
sniffio==1.3.1

# Data Processing
pandas==2.2.3
numpy==2.2.4

# Networking & SSL
certifi==2025.8.3
urllib3==2.5.0
charset-normalizer==3.4.1
idna==3.11

# Type Hints & Utilities
typing_extensions==4.15.0
annotated-types==0.7.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0

# Supabase
#supabase==2.4.2

# Rate Limiting & Caching
redis==4.6.0

# ============================================
# Optional Dependencies (commented out)
# ============================================
# matplotlib==3.10.7          # Plotting (if needed for analytics)
# scikit-learn==1.7.2         # ML models (if needed)
# scipy==1.16.2               # Scientific computing (if needed)
# gunicorn==23.0.0            # Production server (alternative to uvicorn)
